Separate element extraction from elements.json generation
